mail_ru_backuper/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
lxml==5.3.0
pip==22.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==58.1.0
six==1.16.0
urllib3==2.2.3
webdavclient3==3.14.6
